backend/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
pymongo==4.5.0
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==2.3.7
openai==1.3.0
anthropic==0.5.0
requests==2.31.0
gunicorn==21.2.0